Fixes #303: Normalize all paths in the application
In particular:

- project.dir: points to an absolute directory path
- project.buildDir: is relative to project.dir and starts with './'
- project.jar.ouputDir: is relative to project.dir and starts with './'
- PkgDesc.dir: is relative to project.dir and starts with './'

As of this commit, all file and directory paths will be considered
relative to project.dir and they will usually start with './' unless
otherwise stated.

All this normalization is to be able to use `path.join()` with relative
directories (we can because they always start with './') and to be able
to pass directory and file paths freely through they application until
they are finally used in the real filesystem (we can because everything
will be relative to project.dir).

We could also have chosen to use absolute paths always, but that would
be more computationally expensive because we would have to dereference
everything whenever it has to be exported outside of the execution (to
manifest.json, report, console, etc.).

Also, the API is much easier to use and debug with project relative
paths than it is with absolute paths.